+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 ### Added
 
 * Python 3.12 support
+* * `--cache-docker` for `dx build --nextflow`
 
 ### Fixed
 
@@ -48,7 +49,6 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 ### Added
 
 * certifi dxpy dependency
-* `--cache-docker` for `dx build --nextflow`
 
 ### Changed